+++ src/App.tsx
@@ -10,7 +10,8 @@ import { RefreshButton } from "../components/RefreshButton";
 import GlobalContext from "../context/GlobalContext";
 import PlayerCard from "../components/PlayerCard";
 import Logger from "../components/Logger";
-import HelpButton from "../components/HelpButton";
+import HelpView from "../components/HelpView";
+import { FiltersMenu } from "../components/FiltersMenu";
 
 function App() {
   const { data, setData } = useContext(GlobalContext);
@@ -53,10 +54,11 @@ function App() {
             <button onClick={() => setData(null)}>Home</button>
             <RefreshButton />
             <ModifiersMenu />
+            <FiltersMenu />
           </div>
         )}
         <div className="flex flex-row flex-1 gap-x-8 justify-end">
-          <HelpButton />
+          <HelpView />
         </div>
       </header>
       <div className="flex flex-col w-full h-full px-16 pb-8 gap-y-8">
